Move flow queries from LogicDao to UIDao
